refactor: API_BASE_URL 상수 위치 변경
