--- src/apis/core/constants.ts
@@ -1,4 +1,0 @@
-const API_BASE_URL =
-  import.meta.env.VITE_BASE_URL || import.meta.env.STORYBOOK_BASE_URL;
-
-export default API_BASE_URL;
